breaking winmode bindings into separate submodes
Specifically: the layout, split and title bindings.

This is the beginning of a major change in the modes of this config: I'm now going to better document modes with their bindings in the title, so that I don't need to remember every single combination and thus don't get stuck in "Oh, what was the binding for x function?".