Put summary of `git status` in PS1
I'm not sure how I feel about it yet, will give it some time
to try if it's really helpful, if the format is easy to read,
etc.

I've noted that it takes a little time to calculate it all.
I tried to parallelize it with & but it didn't work. It's
not much delay but it might bother some people, let me see
if I'm one of them.

I'm mostly commiting this just to record it, because it took
some time to do; but I might revert it later or simply don't
use the function.