Simplify `git status` in PS1
This sacrifices detail, because I don't know how many files are
untracked, newly tracked, modified and unstaged or modified and
staged, but it's more elegant and simpler to read.  I feel the
number of files doesn't matter much without also know which are
each one of these files and putting their name in prompt would
occupy too much space.

What matters the most here is knowing that there are *some* files
in each of the states mentioned above.  It serves as a headsup
that the command you just did had an effect.  If the user wants
to know which files are in each state, it would be more adequate
to execute `git status` or `git status --short` instead of trying
to represent their output in a condensed format.

I realized that after wanting to also display how many renamed
and staged and renamed and unstaged files there were, but we almost
never rename files on git, so most of the time it would show something
like 0r0, which is just a waste.  A better solution would be to have
R or r show up if there are some files in each state respectively, if
there aren't any r or R, then there are no files what were renamed
since the last commit.  Note, however, that I'm still to implement
this renamed file feature.

I hope I was clear.  I felt like it needed a bit of clarification
on why I prefer this method.  This way myself from the future will
know if he forgets that.