Add alias for `git status --short`
The reason is similar to the one for gitl, alias
for `git log --oneline`. I don't want to get too
used to aliases, but in this case it's simply
too much to type.

Also it's not like `git status --short` is *essential*
for everyday git use, much like `git log --oneline`
is also not essential.