Consider partially staged modified files as unstaged too
When `git add`ing a change to be committed, the user
can use option -p, for example, to only add part of
the change to the commit.  In this case, parts of the
same file are modified and unstaged and others,
modified staged.  This commit makes sure both m and M
flags are up when this happens, as it's expected.