Join `grep` calls and their status vars into oneline
I usually to do multiple commands in oneline, but
this seems like the perfect exception to that.
It makes the whole function fit in the screen,
looks cleaner, doesn't make it harder to read
and honestly it looked like a waste to put the
exit status variables into their own line.