Remove current time from PS1
I'm using tmux now and it has its own clock in the bottom bar.

Also it, one plus of tmux's clock is that it's always up to date,
while the PS1 one shows the time the prompt was generated (the time
the last command was executed). So it can still be useful if you
want to know that, but that is not my particular case.